requirements.txt
Flask==3.0.3
psycopg[binary]==3.2.3
# mysql-connector-python==9.0.0  # Install if using DB_VENDOR=mysql
python-dotenv==1.0.1
